feat: change radiusKm type from Integer to Double in AnalysisSession

+++ backend/comp-intel-core/src/main/java/ua/bkr/monitor/model/AnalysisSession.java
@@ -33,7 +33,7 @@ public class AnalysisSession {
     @Embedded
     private Location location;
 
-    private Integer radiusKm;
+    private Double radiusKm;
 
     @Enumerated(EnumType.STRING)
     private AnalysisStatus status;
